+++ src/components/App.js
@@ -31,7 +31,7 @@ import TermsAndConditions from '../components/Documents/TermsAndConditions';
 import PrivacyPolicy from '../components/Documents/PrivacyPolicy';
 import { APP_LOAD, REDIRECT } from '../constants/actionTypes';
 import { ACCESS_TOKEN, USER } from '../constants/common';
-import { AiofToastContainer } from '../style/common';
+import { AiofToastContainer } from '../style/mui';
 
 
 const mapStateToProps = state => {
