+++ src/components/Home/Banner.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import Grid from '@material-ui/core/Grid';
+import Typography from '@material-ui/core/Typography';
 
-import { H1Alt6, PAlt7 } from '../../style/common';
-import { SquarePaper } from '../../style/mui';
+import { SquarePaper, PAlt7 } from '../../style/mui';
 
 
 const BannerView = props => {
@@ -12,7 +12,7 @@ const BannerView = props => {
       <Grid container>
         <Grid item xs>
           <SquarePaper variant="outlined" square>
-            <H1Alt6>Welcome to {props.appFullName}</H1Alt6>
+            <Typography variant="h1">Welcome to {props.appFullName}</Typography>
             <PAlt7>
               We provide personal finance tools and information to make your life easier
             </PAlt7>
