+++ src/components/PasswordManger.js
@@ -10,8 +10,7 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 
 import { PasswordRuleChecker, ConfirmationPasswordRuleChecker } from './Common/PasswordRuleChecker';
-import { SquarePaper, DefaultRedColor } from '../style/mui';
-import { CoolLink } from '../style/common';
+import { SquarePaper, DefaultRedColor, CoolLink, } from '../style/mui';
 import { success } from './Common/AiofToast';
 import { AiofLoader } from './Common/Loader';
 import { PASSWORD_RESET, PASSWORD_RESET_UNAUTHENTICATED, REDIRECT_HOME, REDIRECT_LOGIN } from '../constants/actionTypes';
