+++ src/components/Property/MortgageCalculator.js
@@ -15,8 +15,7 @@ import { MuiPickersUtilsProvider, KeyboardDatePicker } from '@material-ui/picker
 import { numberWithCommas } from '../Finance/Common';
 import { Line, Bar } from 'react-chartjs-2';
 
-import { AiofLinearProgress, InPaper, SquarePaper, DefaultRedColor, DefaultGreenColor } from '../../style/mui';
-import { ThinText } from '../../style/common';
+import { AiofLinearProgress, InPaper, SquarePaper, DefaultRedColor, DefaultGreenColor, ThinText } from '../../style/mui';
 import { HOUSE_MORTGAGE_CALCULATOR } from '../../constants/actionTypes';
